Version 1.0

+++ src/server/library/helpers/redis_connect.py
@@ -1,6 +1,7 @@
 import sys
 sys.path.append('../server')
 from library.global_variables import REDIS_PORT
+from loguru import logger
 from redis import Redis
 
 
@@ -19,6 +20,7 @@ class RedisConnect:
         get: Retrieves an image from Redis based on the provided image ID.
         stop: Closes the Redis connection.
     """
+    @logger.catch(level='INFO')
     def __init__(self, reader: bool):
         """Initializes the RedisConnect class with a Redis connection and sets the initial image ID.
 
@@ -26,10 +28,14 @@ class RedisConnect:
             reader: A boolean flag indicating whether to clear all data when reading.
         """
         self.connect_redis = Redis(**REDIS_PORT, db=0)
+        if not self.connect_redis.ping():
+            logger.critical('Redis could not connect to the image')
+            sys.exit()
         if reader:
             self.connect_redis.flushall()
         self.id_last_img = 0
 
+    @logger.catch(level='INFO')
     def check_connect(self):
         """Checks the connection status to Redis by pinging.
 
@@ -38,6 +44,7 @@ class RedisConnect:
         """
         return self.connect_redis.ping()
 
+    @logger.catch(level='INFO')
     def set(self, img, id_img=None):
         """Sets an image in Redis with or without specifying an image ID.
 
@@ -55,6 +62,7 @@ class RedisConnect:
             self.id_last_img += 1
             return str(self.id_last_img - 1)
 
+    @logger.catch(level='INFO')
     def get(self, id_img):
         """Retrieves an image from Redis based on the provided image ID.
 
@@ -67,6 +75,7 @@ class RedisConnect:
         img = self.connect_redis.get(id_img)
         return img
 
+    @logger.catch(level='INFO')
     def stop(self):
         """Closes the Redis connection.
         """
